Add middleware options file with content.
